+++ apps/restconf/restconf_methods_post.h
@@ -42,14 +42,13 @@
  * Prototypes
  */
 int api_data_post(clicon_handle h, FCGX_Request *r, char *api_path,
-		  cvec *pcvec, int pi, 
+		  int pi,
 		  cvec *qvec, char *data,
 		  int pretty,
 		  restconf_media media_out);
 
-int api_operations_post(clicon_handle h, FCGX_Request *r, 
-			char *path,
-			cvec *pcvec, int pi, cvec *qvec, char *data,
+int api_operations_post(clicon_handle h, FCGX_Request *r, char *api_path,
+			int pi,	cvec *qvec, char *data,
 			int pretty, restconf_media media_out);