+++ lib/clixon/clixon_path.h
@@ -79,8 +79,7 @@ int xml_yang_root(cxobj *x, cxobj **xr);
 int yang2api_path_fmt(yang_stmt *ys, int inclkey, char **api_path_fmt);
 int api_path_fmt2api_path(char *api_path_fmt, cvec *cvv, char **api_path);
 int api_path_fmt2xpath(char *api_path_fmt, cvec *cvv, char **xpath);
-int api_path2xpath_cvv(cvec *api_path, int offset, yang_stmt *yspec, cbuf *xpath, cvec **nsc, cxobj **xerr);
-int api_path2xpath(char *api_path, yang_stmt *yspec, char **xpath, cvec **nsc);
+int api_path2xpath(char *api_path, yang_stmt *yspec, char **xpath, cvec **nsc, cxobj **xerr);
 int api_path2xml(char *api_path, yang_stmt *yspec, cxobj *xtop, 
 		 yang_class nodeclass, int strict,
 		 cxobj **xpathp, yang_stmt **ypathp, cxobj **xerr);